+++ module/index.js
@@ -11,6 +11,7 @@ import { acceptNumber, rejectNumber } from './configs/empty';
 import config_number from './configs/number';
 import config_text from './configs/text';
 import config_function from './configs/function';
+import config_boolean from './configs/boolean';
 
 var configs = {
   array: config_array,
@@ -21,7 +22,9 @@ var configs = {
   rejectNumber: rejectNumber,
   number: config_number,
   text: config_text,
-  function: config_function
+  string: config_text,
+  function: config_function,
+  boolean: config_boolean
 };
 
 /**