+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+<a name="1.3.0"></a>
+# [1.3.0](https://github.com/InlineManual/coerce/compare/v1.2.0...v1.3.0) (2016-10-25)
+
+
+### Features
+
+* add `boolean` config ([6fbe9dd](https://github.com/InlineManual/coerce/commit/6fbe9dd))
+* add alias `string` for `text` ([335879e](https://github.com/InlineManual/coerce/commit/335879e))
+
+
+
 <a name="1.2.0"></a>
 # [1.2.0](https://github.com/InlineManual/coerce/compare/v1.1.0...v1.2.0) (2016-09-26)